LayoutTests/platform/ios-sim-deprecated/http/tests/loading/nested_bad_objects-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
 PASS - nested image objects with bad mimetype do not cause a crash.
